+++ src/convert/styles/css/fontStyle.ts
@@ -1,4 +1,8 @@
-import { VscodeEnvLanguage, getEnvLanguage, getPaperbackWriterConfiguration } from "../../../vscode"
+import {
+  VscodeEnvLanguage,
+  getEnvLanguage,
+  getPaperbackWriterConfiguration
+} from "../../../util"
 import { FontQuery } from "../styleTagBuilder"
 
 /**
