+++ src/convert/styles/styleTagBuilder.ts
@@ -1,6 +1,14 @@
 import path from "path"
 import * as vscode from 'vscode'
-import { VscodeEnvLanguage, getHomeDirPath, getPaperbackWriterConfiguration, getUri, getVscodeUri, getWorkspaceFolder, showMessage } from "../../vscode"
+import {
+  VscodeEnvLanguage,
+  getHomeDirPath,
+  getPaperbackWriterConfiguration,
+  getUri,
+  getVscodeUri,
+  getWorkspaceFolder,
+  showMessage
+} from "../../util"
 import { remedyCss } from "./css/remedyCss"
 import { FontSetId, getFontFamily } from "./css/fontStyle"
 import { blockquoteCss } from "./css/blockquoteCss"
