+++ src/util/index.ts
@@ -1,0 +1,9 @@
+export { 
+  isExistsPath,
+  mkdir,
+  isExistsDir,
+  deleteFile,
+  getOutputPathName,
+  hasExtension
+ } from './util'
+export * from './vscode'
