+++ src/util/util.ts
@@ -1,6 +1,11 @@
 import fs from 'fs'
 import path from "path"
-import { getNls, getPaperbackWriterConfiguration, getWorkspaceFolder, showMessage } from './vscode'
+import {
+  getNls,
+  getPaperbackWriterConfiguration,
+  getWorkspaceFolder,
+  showMessage
+} from './'
 import { mkdirp } from "mkdirp"
 import os from "os"
 import * as vscode from 'vscode'
