+++ src/util/vscode/vscodeSettingUtil.ts
@@ -1,9 +1,9 @@
 import * as vscode from 'vscode'
-import { PdfOrientation } from '../export/pdf/exportPdf'
-import { PaperbackWriterOutputType } from '../paperbackWriter'
+import { PdfOrientation } from '../../export/pdf/exportPdf'
+import { PaperbackWriterOutputType } from '../../paperbackWriter'
 import { PaperFormat } from 'puppeteer'
-import { HeaderFooterItems } from '../export/pdf/pdfHeaderFooterUtil'
-import { CodeThemeName } from '../convert/markdown/customRenderer/customRendererCodeBlockThemeCSSGenerator'
+import { HeaderFooterItems } from '../../export/pdf/pdfHeaderFooterUtil'
+import { CodeThemeName } from '../../convert/markdown/customRenderer/customRendererCodeBlockThemeCSSGenerator'
 
 type HeaderFooterFontSize = "50%" | "60%" | "70%" | "80%" | "90%"
